requirements.txt
debugpy
aiohttp
python-dotenv
